refactor(capability): split verifiedCapabilityInvocation behind its public facade
verifiedCapabilityInvocation.ts 1257 -> 65 and leaves the filesize allowlist. It has sat
there on main for a long time and is unrelated to the M1 agent-host work.

The file held three separate responsibilities. Each becomes a module; the original path stays
as the named public surface because 16 callers import a mix of symbols from it, and rewiring
them into two or three imports each would be worse than one facade:
- verifiedCapabilityInvocationRuntime.ts (357): invocation types, policy revisions, canonical
  hashing/freezing, mint state, execution-target resolution, revalidation.
- verifiedCapabilityInvocationSessionFactories.ts (321): MCP leased-session and internal
  bearer request parsing/evidence plus their mint factories.
- verifiedCapabilityInvocationRendererFactories.ts (657): renderer surface/snapshot factories
  and the renderer document, canvas, timeline, asset and export target policy factories.
No per-record seam was created — the cross-module calls sit at invocation/setup and
execution-validation boundaries.

Two sibling splits were prepared and then deliberately dropped from this branch:
- projectAgentReducer.ts 936 -> 759 regresses the hot path when ported to main on its own.
  The reducer perf assertion measures 18847ms on pristine main and 32329ms with the split,
  against an 8000ms budget — 1.7x worse. It only helped on a branch where the execution
  coordinator was split alongside it, so the win came from the coordinator half. Revisit the
  two together, and re-measure.
- productionRunService.ts 816 -> 668 is blocked by the root-cause contract gate, which
  requires a docs/fixes/*.root-cause.json for any change under electron/productionRun/. That
  schema is shaped for corrective work (symptom, direct_cause, class_root, prevention); a
  pure structural move has none of those, and filling the fields in anyway would be fiction.
  Left for a decision on whether the gate should be able to express a non-corrective change.

Verified: pnpm run typecheck EXIT=0; npx vitest run electron/capabilityCore/ +
electron/productionRun/ EXIT=0 with 55 files / 359 tests passed; check-file-sizes EXIT=0.
